tank_project/requirements.txt
# Dépendances Projet Tank Arena

# Vision & Traitement Image
numpy>=1.20.0
opencv-python>=4.5.0
pyrealsense2>=2.50.0

# Visualisation
pygame>=2.1.0

# Traitement Signal & Math
scipy>=1.7.0

# Configuration
pyyaml>=5.4.0

# Utilitaires
matplotlib>=3.3.0
